+++ mobile/android/fenix/app/src/androidTest/java/org/mozilla/fenix/helpers/AppAndSystemHelper.kt
@@ -312,27 +312,25 @@ object AppAndSystemHelper {
         }
     }
 
-    fun assertAppWithPackageNameOpens(appPackageName: String) {
+    fun assertExternalAppOpens(appPackageName: String) {
         if (isPackageInstalled(appPackageName)) {
             try {
-                Log.i(TAG, "assertAppWithPackageNameOpens: Trying to check the intent sent.")
+                Log.i(TAG, "assertExternalAppOpens: Trying to check the intent sent.")
                 intended(toPackage(appPackageName))
-                Log.i(TAG, "assertAppWithPackageNameOpens: Matched open intent to $appPackageName.")
+                Log.i(TAG, "assertExternalAppOpens: Matched open intent to $appPackageName.")
             } catch (e: AssertionFailedError) {
-                Log.i(TAG, "assertAppWithPackageNameOpens: Intent match failure. ${e.message}")
+                Log.i(TAG, "assertExternalAppOpens: Intent match failure. ${e.message}")
             } finally {
-                if (appPackageName != packageName) {
-                    // Stop the app from running in the background
-                    forceCloseApp(appPackageName)
-                }
+                // Stop the app from running in the background
+                forceCloseApp(appPackageName)
             }
         } else {
-            Log.i(TAG, "assertAppWithPackageNameOpens: Trying to verify the \"Could not open file\" message.")
+            Log.i(TAG, "assertExternalAppOpens: Trying to verify the \"Could not open file\" message.")
             mDevice.waitNotNull(
                 Until.findObject(By.text("Could not open file")),
                 waitingTime,
             )
-            Log.i(TAG, "assertAppWithPackageNameOpens: Verified \"Could not open file\" message")
+            Log.i(TAG, "assertExternalAppOpens: Verified \"Could not open file\" message")
         }
     }